+++ aula-3/src/main/java/org/acelera/saopaulo/soccer/Posicao.java
@@ -1,4 +1,4 @@
-package org.acelera.saopaulo.futebol;
+package org.acelera.saopaulo.soccer;
 
 public enum Posicao {
