+++ aula-3/src/main/java/org/acelera/saopaulo/soccer/controller/JogadorController.java
@@ -1,7 +1,7 @@
-package org.acelera.saopaulo.futebol.controller;
+package org.acelera.saopaulo.soccer.controller;
 
-import org.acelera.saopaulo.futebol.dto.JogadorDto;
-import org.acelera.saopaulo.futebol.service.JogadorService;
+import org.acelera.saopaulo.soccer.dto.JogadorDto;
+import org.acelera.saopaulo.soccer.service.JogadorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
