+++ aula-3/src/main/java/org/acelera/saopaulo/soccer/repository/TimeRepository.java
@@ -1,6 +1,6 @@
-package org.acelera.saopaulo.futebol.repository;
+package org.acelera.saopaulo.soccer.repository;
 
-import org.acelera.saopaulo.futebol.entity.Time;
+import org.acelera.saopaulo.soccer.entity.Time;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
