+++ aula-3/src/main/java/org/acelera/saopaulo/soccer/service/TimeService.java
@@ -1,12 +1,12 @@
-package org.acelera.saopaulo.futebol.service;
+package org.acelera.saopaulo.soccer.service;
 
-import org.acelera.saopaulo.futebol.Posicao;
-import org.acelera.saopaulo.futebol.builder.JogadorDtoBuilder;
-import org.acelera.saopaulo.futebol.builder.TimeDtoBuilder;
-import org.acelera.saopaulo.futebol.dto.JogadorDto;
-import org.acelera.saopaulo.futebol.dto.TimeDto;
-import org.acelera.saopaulo.futebol.entity.Time;
-import org.acelera.saopaulo.futebol.repository.TimeRepository;
+import org.acelera.saopaulo.soccer.Posicao;
+import org.acelera.saopaulo.soccer.builder.JogadorDtoBuilder;
+import org.acelera.saopaulo.soccer.builder.TimeDtoBuilder;
+import org.acelera.saopaulo.soccer.dto.JogadorDto;
+import org.acelera.saopaulo.soccer.dto.TimeDto;
+import org.acelera.saopaulo.soccer.entity.Time;
+import org.acelera.saopaulo.soccer.repository.TimeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
