+++ aula-3/src/test/java/org/acelera/saopaulo/AppTest.java
@@ -1,23 +1,16 @@
 package org.acelera.saopaulo;
 
 
-import com.sun.deploy.net.HttpResponse;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-
-import java.io.IOException;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 
-public class AppTest {
-
+public class AppTest 
+{
     @Test
     public void shouldAnswerWithTrue()
     {
         assertTrue( true );
     }
-
-
-
 }
